refactor: extract native playback fakes
